requirements.txt
# Telegram bot framework
python-telegram-bot==21.4

# HTTP requests
requests==2.32.3
httpx==0.27.0

# Core utilities
urllib3==2.2.2

# Date and time handling
pytz==2024.1

# For data parsing / regex / HTML escaping (standard libs are fine but listed for safety)
html5lib==1.1

# Optional — only if your system doesn’t have these standard libs by default
# (Usually not needed, but harmless)
difflib
re
threading
asyncio
